# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Application de sondage de dates type Doodle avec authentification Google OAuth et email/password.",
        "title": "Doodle Clone API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Stéphane LE MINH NHUT",
            "url": "https://github.com",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/polls/{id}/export/csv": {
            "get": {
                "description": "Génère un fichier CSV des votes pour analyse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Exporter en CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID du sondage",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/polls/{id}/export/ics": {
            "get": {
                "description": "Génère un fichier calendrier (Google Calendar, Outlook)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/calendar"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Exporter en ICS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID du sondage",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/polls/{id}/export/pdf": {
            "get": {
                "description": "Génère un fichier PDF du sondage avec les résultats",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Exporter en PDF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID du sondage",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}